refactor[tcg]: separate files for each character
