+++ tcg/characters/index.ts
@@ -1,0 +1,14 @@
+/**
+ * TCG character roster. Definitions live under `tcg/characters/` (one file per character);
+ * {@link ./roster.ts} aggregates them into {@link CHARACTERS}.
+ */
+
+export * from './roster';
+export * from './shared';
+export * from './kaitlin';
+export * from './venfei';
+export * from './ei';
+export * from './silverwolf';
+export * from './sparkle';
+export * from './electro';
+export * from './mystic';
