+++ tcg/characters/shared.ts
@@ -1,0 +1,110 @@
+import type { CharacterTextColors } from '../textTheme';
+
+export const FAIRY_TEXT_COLORS: CharacterTextColors = {
+  nameFill: '#FFF6F9',
+  nameStroke: '#3A1323',
+  hpLabelFill: '#FFD8EC',
+  hpLabelStroke: '#3A1323',
+  hpValueFill: '#FFFFFF',
+  hpValueStroke: '#2B1020',
+  titleFill: '#FFE5F2',
+  titleStroke: '#2E1320',
+  titleDescFill: '#FFF9FD',
+  titleDescStroke: '#23111A',
+  skillNameFill: '#FFE0F0',
+  skillNameStroke: '#2E1320',
+  skillDamageFill: '#FFFFFF',
+  skillDamageStroke: '#2B1020',
+  skillCostFill: '#FFEFFF',
+  skillCostStroke: '#2B1636',
+  skillDescFill: '#FFFFFF',
+  skillDescStroke: '#221824',
+  abilityNameFill: '#FFE0F0',
+  abilityNameStroke: '#2E1320',
+  abilityDescFill: '#FFFFFF',
+  abilityDescStroke: '#221824',
+};
+
+export const FAIRY_ABILITY_PANEL_COLOR = '#D5ABB2';
+
+export const QUANTUM_TEXT_COLORS: CharacterTextColors = {
+  nameFill: '#EDE2FF',
+  nameStroke: '#1A123A',
+  hpLabelFill: '#D7CCFF',
+  hpLabelStroke: '#1A123A',
+  hpValueFill: '#F8F3FF',
+  hpValueStroke: '#16122D',
+  titleFill: '#E5D8FF',
+  titleStroke: '#1B1433',
+  titleDescFill: '#F8F3FF',
+  titleDescStroke: '#1A1730',
+  skillNameFill: '#E3D5FF',
+  skillNameStroke: '#1B1433',
+  skillDamageFill: '#FDFBFF',
+  skillDamageStroke: '#16122D',
+  skillCostFill: '#CDE8FF',
+  skillCostStroke: '#18223D',
+  skillDescFill: '#F9F5FF',
+  skillDescStroke: '#1A1730',
+  abilityNameFill: '#E3D5FF',
+  abilityNameStroke: '#1B1433',
+  abilityDescFill: '#F9F5FF',
+  abilityDescStroke: '#1A1730',
+};
+
+export const QUANTUM_ABILITY_PANEL_COLOR = '#5539CC';
+
+/** Purple-violet card text (pairs with cyan Electro backgrounds). */
+export const ELECTRO_TEXT_COLORS: CharacterTextColors = {
+  nameFill: '#F4E9FF',
+  nameStroke: '#2A1445',
+  hpLabelFill: '#E4D4FF',
+  hpLabelStroke: '#2A1445',
+  hpValueFill: '#FCF7FF',
+  hpValueStroke: '#1E0F35',
+  titleFill: '#ECD9FF',
+  titleStroke: '#261742',
+  titleDescFill: '#FAF4FF',
+  titleDescStroke: '#1E1538',
+  skillNameFill: '#E8D6FF',
+  skillNameStroke: '#2A1A4A',
+  skillDamageFill: '#FFF9FF',
+  skillDamageStroke: '#1A0F32',
+  skillCostFill: '#DCC4FF',
+  skillCostStroke: '#2D1F52',
+  skillDescFill: '#F7F0FF',
+  skillDescStroke: '#22183C',
+  abilityNameFill: '#E6D2FF',
+  abilityNameStroke: '#2A1A4A',
+  abilityDescFill: '#F7F0FF',
+  abilityDescStroke: '#22183C',
+};
+
+export const ELECTRO_ABILITY_PANEL_COLOR = '#39AACC';
+
+export const ANEMO_TEXT_COLORS: CharacterTextColors = {
+  nameFill: '#E8FFF4',
+  nameStroke: '#123A2A',
+  hpLabelFill: '#C8F5E0',
+  hpLabelStroke: '#123A2A',
+  hpValueFill: '#F4FFFA',
+  hpValueStroke: '#0E2E22',
+  titleFill: '#D4F5E4',
+  titleStroke: '#143528',
+  titleDescFill: '#F0FFF8',
+  titleDescStroke: '#102820',
+  skillNameFill: '#C8EDDA',
+  skillNameStroke: '#143528',
+  skillDamageFill: '#FFFFFF',
+  skillDamageStroke: '#0E2E22',
+  skillCostFill: '#D8FFE8',
+  skillCostStroke: '#1A3D30',
+  skillDescFill: '#F4FFFA',
+  skillDescStroke: '#102820',
+  abilityNameFill: '#C8EDDA',
+  abilityNameStroke: '#143528',
+  abilityDescFill: '#F4FFFA',
+  abilityDescStroke: '#102820',
+};
+
+export const ANEMO_ABILITY_PANEL_COLOR = '#5FBF8A';
